small bugfix

+++ core/data_handler/includes/modules/write/calendar_raids_attendees/pdh_w_calendar_raids_attendees.class.php
@@ -62,7 +62,7 @@ if(!class_exists('pdh_w_calendar_raids_attendees')){
 					'member_id'				=> $memberid,
 					'calendar_events_id'	=> $eventid,
 					'timestamp_signup'		=> $this->time->time,
-					'raidgroup'				=> ((int)$raidgroup > 0) ? $raidgroup : array_shift($this->pdh->get('raid_groups_members', 'memberships', array($memberid))),
+					'raidgroup'				=> ((int)$raidgroup > 0) ? $raidgroup : (int)array_shift($this->pdh->get('raid_groups_members', 'memberships', array($memberid))),
 					'random_value'			=> $rand_value,
 					'member_role'			=> $memberrole,
 					'signup_status'			=> $signupstatus,
